+++ killtracker/__init__.py
@@ -1,6 +1,6 @@
 default_app_config = "killtracker.apps.KillmailsConfig"
 
-__version__ = "0.3.0"
+__version__ = "0.3.0a1"
 __title__ = "Killtracker"
 
 USER_AGENT_TEXT = f"aa-killtracker v{__version__}"