+++ killtracker/admin.py
@@ -72,6 +72,7 @@ class TrackerAdmin(admin.ModelAdmin):
         "webhook",
         "identify_fleets",
         "_clauses",
+        "_color",
     )
     list_filter = (
         "is_enabled",
@@ -79,6 +80,12 @@ class TrackerAdmin(admin.ModelAdmin):
         ("webhook", admin.RelatedOnlyFieldListFilter),
     )
     ordering = ("name",)
+
+    def _color(self, obj):
+        html = (
+            f'<input type="color" value="{obj.color}" disabled>' if obj.color else "-"
+        )
+        return mark_safe(html)
 
     def _clauses(self, obj):
         clauses = list()
@@ -127,7 +134,10 @@ class TrackerAdmin(admin.ModelAdmin):
     def _append_field_to_clauses(self, clauses, field, text):
         clauses.append(f"{field_nice_display(field)} = {text}")
 
-    actions = ["run_test_killmail"]
+    actions = ["run_test_killmail", "reset_color"]
+
+    def reset_color(self, request, queryset):
+        queryset.update(color=None)
 
     def run_test_killmail(self, request, queryset):
         if "apply" in request.POST:
@@ -199,7 +209,7 @@ class TrackerAdmin(admin.ModelAdmin):
     )
 
     fieldsets = (
-        (None, {"fields": ("name", "description", "is_enabled")}),
+        (None, {"fields": ("name", "description", "is_enabled", "color")}),
         (
             "Discord Configuration",
             {